+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - [#82](https://github.com/projectapheleia/avl/issues/82) Vanilla Template : Issues with sync_reset and ticker
  - [#88](https://github.com/projectapheleia/avl/issues/88) Coverbin(): getstddev asserts if _count_ <=1
  - [#89](https://github.com/projectapheleia/avl/issues/89) IndexedScoreboard(): Ordering of checks can be improved by threading the before and after ports
+ - [#90](https://github.com/projectapheleia/avl/issues/90) Object(): manipulation of _auto_random_ can cause false constraint fails
 
 ## [v0.5.1] - 2026-03-05